+++ tokenizer/src/tokenizers.rs
@@ -2,8 +2,10 @@
 //
 // See LICENSE in the repository root for license terms.
 
+mod jieba;
 mod unicode;
 mod whitespace;
 
+pub(crate) use jieba::*;
 pub(crate) use unicode::*;
 pub use whitespace::*;